backend/requirements-dev.txt
# Development dependencies for contract tooling
# Pinned versions for deterministic builds
datamodel-code-generator==0.25.0
pydantic>=2.0.0,<3.0.0

# Contract-first enforcement dependencies
schemathesis>=3.19.0  # Spec-driven API testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # For async client testing
pyyaml>=6.0  # For config parsing
fastapi>=0.100.0
uvicorn>=0.23.0